pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-ai"
version = "0.1.0"
description = "An open-source autonomous AI assistant that actually does things"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Abduvohid" }
]
keywords = ["ai", "assistant", "autonomous", "agent", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "litellm>=1.55,<2",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "typer>=0.15",
    "rich>=13.7",
    "structlog>=24.4",
    "pyyaml>=6.0",
    "httpx>=0.28,<1",
    "python-dotenv>=1.0",
    "apscheduler>=3.10,<4",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "aiosqlite>=0.20",
    "duckduckgo-search>=7.0,<8",
    "edge-tts>=7.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<10",
    "pytest-asyncio>=1.0,<2",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.11",
    "pre-commit>=3.8",
]
memory = [
    "chromadb>=1.0,<2",
    "sentence-transformers>=5.0",
]
desktop = [
    "pyautogui>=0.9.54",
    "Pillow>=10.0",
    "pyperclip>=1.8",
    "pygetwindow>=0.0.9; sys_platform == 'win32'",
]
claude-sdk = [
    "claude-agent-sdk>=0.1.0",
]
voice = [
    "faster-whisper>=1.0",
]
all = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "aiogram>=3.10",
    "aiosqlite>=0.20",
    "chromadb>=1.0,<2",
    "sentence-transformers>=5.0",
    "apscheduler>=3.10,<4",
    "playwright>=1.45",
    "websockets>=14.0",
    "edge-tts>=7.0",
    "pyautogui>=0.9.54",
    "Pillow>=10.0",
    "pyperclip>=1.8",
    "pygetwindow>=0.0.9; sys_platform == 'win32'",
    "claude-agent-sdk>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/OWNER/agent"
Documentation = "https://github.com/OWNER/agent/tree/main/docs"
Repository = "https://github.com/OWNER/agent"
Issues = "https://github.com/OWNER/agent/issues"
Changelog = "https://github.com/OWNER/agent/blob/main/CHANGELOG.md"

[project.scripts]
agent = "agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    # chromadb creates an internal event loop that it doesn't close
    "ignore:unclosed event loop:ResourceWarning",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "ASYNC"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
